Copy Maptime CoC from meetup.com/Maptime-NYC
In case MaptimeNYC.github.io/MaptimeNYC becomes a central tool for
communicating and organizing, it is important that reference docs are
easily accessible.  This commit copies the Code of Conduct that is
posted on our Meetup page, using Markup to match the document design.

See:  meetup.com/Maptime-NYC/pages/19670921/Maptime_Code_of_Conduct/

Credit to Chris Henrick for taking the time to make the CoC prominent
on the Meetup page to ensure proper guidelines for participation in the
community.